starting to debug a m_bytes_used var for DA

+++ src/umpire/DeviceAllocator.hpp
@@ -49,7 +49,7 @@ class DeviceAllocator {
   unsigned int* m_counter;
 
   size_t m_id;
-  size_t m_bytes_used;
+  unsigned int m_bytes_used;
   size_t m_size;
   bool m_initialized;
   bool m_child;
